Start of new endpoint to return the percentages from the purchases summary
