+++ api/serializers/__init__.py
@@ -34,7 +34,12 @@ from .blogpost import BlogPostSerializer  # noqa: F401
 from .password import PasswordSerializer  # noqa: F401
 from .managerinvitation import ManagerInvitationSerializer  # noqa: F401
 from .teledeclaration import ShortTeledeclarationSerializer  # noqa: F401
-from .purchase import PurchaseSerializer, PurchaseSummarySerializer, PurchaseExportSerializer  # noqa: F401
+from .purchase import (  # noqa: F401
+    PurchaseSerializer,
+    PurchaseSummarySerializer,
+    PurchasePercentageSummarySerializer,
+    PurchaseExportSerializer,
+)
 from .reservationexpe import ReservationExpeSerializer  # noqa: F401
 from .vegetarianexpe import VegetarianExpeSerializer  # noqa: F401
 from .message import MessageSerializer  # noqa: F401
