+++ api/tests/test_purchases.py
@@ -559,6 +559,122 @@ class TestPurchaseApi(APITestCase):
         self.assertEquals(body["results"][1]["year"], 2021)
         self.assertEquals(body["results"][1]["valueTotalHt"], 450)
 
+    def test_get_public_purchases_summary(self):
+        """
+        Return percentages from purchase data for the given year and canteen
+        """
+        canteen = CanteenFactory.create()
+        year = 2024
+
+        # bio percent, ignore lesser labels
+        PurchaseFactory.create(
+            canteen=canteen,
+            date="2024-01-01",
+            characteristics=[Purchase.Characteristic.BIO, Purchase.Characteristic.LABEL_ROUGE],
+            family=Purchase.Family.VIANDES_VOLAILLES,
+            price_ht=10,
+        )
+        # sustainable percent, meat egalim
+        PurchaseFactory.create(
+            canteen=canteen,
+            date="2024-01-01",
+            characteristics=[Purchase.Characteristic.LABEL_ROUGE],
+            family=Purchase.Family.VIANDES_VOLAILLES,
+            price_ht=10,
+        )
+        # externalities percent, meat egalim, meat france
+        PurchaseFactory.create(
+            canteen=canteen,
+            date="2024-01-01",
+            characteristics=[Purchase.Characteristic.EXTERNALITES, Purchase.Characteristic.FRANCE],
+            family=Purchase.Family.VIANDES_VOLAILLES,
+            price_ht=10,
+        )
+        # egalim others, fish egalim
+        PurchaseFactory.create(
+            canteen=canteen,
+            date="2024-01-01",
+            characteristics=[Purchase.Characteristic.PECHE_DURABLE],
+            family=Purchase.Family.PRODUITS_DE_LA_MER,
+            price_ht=10,
+        )
+        # meat france (local and short_distribution not included?)
+        PurchaseFactory.create(
+            canteen=canteen,
+            date="2024-12-31",
+            characteristics=[Purchase.Characteristic.FRANCE],
+            family=Purchase.Family.VIANDES_VOLAILLES,
+            price_ht=10,
+        )
+        # fish non egalim
+        PurchaseFactory.create(
+            canteen=canteen,
+            date="2024-12-31",
+            characteristics=[Purchase.Characteristic.FRANCE],
+            family=Purchase.Family.PRODUITS_DE_LA_MER,
+            price_ht=10,
+        )
+        # add misc purchase to have nice round total of 100 HT
+        PurchaseFactory.create(
+            canteen=canteen,
+            date="2024-12-31",
+            characteristics=[],
+            family=Purchase.Family.AUTRES,
+            price_ht=40,
+        )
+
+        # create purchase outside of requested year to check filtering
+        PurchaseFactory.create(
+            canteen=canteen,
+            date="2023-12-31",
+            characteristics=[Purchase.Characteristic.BIO],
+            family=Purchase.Family.VIANDES_VOLAILLES,
+            price_ht=999999,
+        )
+
+        response = self.client.get(
+            reverse("canteen_purchases_percentage_summary", kwargs={"canteen_pk": canteen.id}), {"year": year}
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        body = response.json()
+
+        self.assertEqual(body["percentageValueBioHt"], 0.1)
+        self.assertEqual(body["percentageValueSustainableHt"], 0.1)
+        self.assertEqual(body["percentageValueExternalityPerformanceHt"], 0.1)
+        self.assertEqual(body["percentageValueEgalimOthersHt"], 0.1)
+        # 30 HT, meat total is 40.
+        self.assertEqual(body["percentageValueMeatPoultryEgalimHt"], 0.75)
+        self.assertEqual(body["percentageValueMeatPoultryFranceHt"], 0.5)
+        self.assertEqual(body["percentageValueFishEgalimHt"], 0.5)
+
+    def test_cannot_get_redacted_purchases_summary(self):
+        """
+        If the canteen has redacted the year return a 400
+        """
+        pass
+
+    def test_no_purchases_for_public_summary(self):
+        """
+        If the canteen doesn't have purchases for the year requested return a 400
+        """
+        pass
+
+    @authenticate
+    def test_get_last_purchase_date_in_public_summary_if_manager(self):
+        """
+        The purchases summary should return the last purchase date if the user
+        is the manager of the canteen
+        """
+        pass
+
+    @authenticate
+    def test_dont_get_last_purchase_date_in_public_summary_if_not_manager(self):
+        """
+        The purchases summary should not return the last purchase date if the user
+        is not the manager of the canteen, even if authenticated
+        """
+        pass
+
     @authenticate
     def test_delete_purchase(self):
         """
@@ -924,6 +1040,7 @@ class TestPurchaseApi(APITestCase):
             family=Purchase.Family.BOISSONS,
             characteristics=[Purchase.Characteristic.AOCAOP],
         )
+        # TODO: would be nice to double check the AOCAOP IGP STG aggregation vs other labels
         PurchaseFactory.create(
             canteen=canteen_site,
             date="2021-12-31",
@@ -954,6 +1071,7 @@ class TestPurchaseApi(APITestCase):
         diag_site = Diagnostic.objects.get(year=year, canteen=canteen_site)
         self.assertIn(diag_site.id, results)
         self.assertEqual(diag_site.value_total_ht, 200)
+        # TODO: would be nice to test the aggregation for a simple value (e.g. value_sustainable_ht)
         self.assertEqual(diag_site.value_boissons_aocaop_igp_stg, 50)
         self.assertEqual(diag_site.value_boulangerie_non_egalim, 150)
         self.assertEqual(diag_site.value_autres_aocaop_igp_stg, 0)
