+++ api/views/purchase.py
@@ -12,7 +12,12 @@ from django.db.models.functions import ExtractYear
 from django.http import JsonResponse
 from django_filters import rest_framework as django_filters
 from api.permissions import IsLinkedCanteenManager, IsCanteenManager, IsAuthenticated
-from api.serializers import PurchaseSerializer, PurchaseSummarySerializer, PurchaseExportSerializer
+from api.serializers import (
+    PurchaseSerializer,
+    PurchaseSummarySerializer,
+    PurchasePercentageSummarySerializer,
+    PurchaseExportSerializer,
+)
 from data.models import Purchase, Canteen, Diagnostic
 from .utils import MaCantineOrderingFilter, UnaccentSearchFilter
 from collections import OrderedDict
@@ -178,13 +183,32 @@ class CanteenPurchasesSummaryView(APIView):
         canteen_id = kwargs.get("canteen_pk")
         canteen = self._get_canteen(canteen_id, self.request)
         year = request.query_params.get("year")
-        return Response(canteen_summary_for_year(canteen, year) if year else canteen_summary(canteen))
+        data = canteen_summary_for_year(canteen, year) if year else canteen_summary(canteen)
+        return Response(PurchaseSummarySerializer(data).data if year else data)
 
     def _get_canteen(self, canteen_id, request):
         try:
             canteen = Canteen.objects.get(pk=canteen_id)
             if not IsCanteenManager().has_object_permission(request, self, canteen):
                 raise PermissionDenied()
+            return canteen
+        except Canteen.DoesNotExist as e:
+            raise NotFound() from e
+
+
+class CanteenPurchasesPercentageSummaryView(APIView):
+    def get(self, request, *args, **kwargs):
+        canteen_id = kwargs.get("canteen_pk")
+        canteen = self._get_canteen(canteen_id, self.request)
+        year = request.query_params.get("year")
+        if not year:
+            raise BadRequest("year is required")
+        data = canteen_summary_for_year(canteen, year)
+        return Response(PurchasePercentageSummarySerializer(data).data)
+
+    def _get_canteen(self, canteen_id, request):
+        try:
+            canteen = Canteen.objects.get(pk=canteen_id)
             return canteen
         except Canteen.DoesNotExist as e:
             raise NotFound() from e
@@ -199,7 +223,7 @@ def canteen_summary_for_year(canteen, year):
     complete_diag_data(purchases, data)
     misc_totals(purchases, data)
 
-    return PurchaseSummarySerializer(data).data
+    return data
 
 
 def canteen_summary(canteen):
@@ -220,10 +244,21 @@ def canteen_summary(canteen):
 
 
 # the order of EGALIM_LABELS is significant - determines which labels trump others when aggregating purchases
-EGALIM_LABELS = [
+DIAGNOSTIC_EGALIM_LABELS = [
     "BIO",
     "LABEL_ROUGE",
     "AOCAOP_IGP_STG",
+    "HVE",
+    "PECHE_DURABLE",
+    "RUP",
+    "COMMERCE_EQUITABLE",
+    "FERMIER",
+    "EXTERNALITES",
+    "PERFORMANCE",
+]
+PURCHASE_EGALIM_LABELS = [
+    "BIO",
+    "LABEL_ROUGE",
     "AOCAOP",
     "IGP",
     "STG",
@@ -239,6 +274,7 @@ EGALIM_LABELS = [
 
 
 def simple_diag_data(purchases, data):
+    # TODO: is CONVERSION_BIO used?
     bio_filter = Q(characteristics__contains=[Purchase.Characteristic.BIO]) | Q(
         characteristics__contains=[Purchase.Characteristic.CONVERSION_BIO]
     )
@@ -297,7 +333,7 @@ def complete_diag_data(purchases, data):
 
     for family in families:
         purchase_family = purchases.filter(family=family)
-        for label in EGALIM_LABELS:
+        for label in DIAGNOSTIC_EGALIM_LABELS:
             if label == "AOCAOP_IGP_STG":
                 fam_label = purchase_family.filter(
                     Q(characteristics__contains=[Purchase.Characteristic.AOCAOP])
@@ -341,7 +377,7 @@ def misc_totals(purchases, data):
     )
     data["value_meat_poultry_ht"] = meat_poultry_purchases.aggregate(total=Sum("price_ht"))["total"] or 0
 
-    meat_poultry_egalim = meat_poultry_purchases.filter(characteristics__overlap=EGALIM_LABELS)
+    meat_poultry_egalim = meat_poultry_purchases.filter(characteristics__overlap=PURCHASE_EGALIM_LABELS)
     data["value_meat_poultry_egalim_ht"] = meat_poultry_egalim.aggregate(total=Sum("price_ht"))["total"] or 0
 
     meat_poultry_france = meat_poultry_purchases.filter(
@@ -356,7 +392,7 @@ def misc_totals(purchases, data):
     )
     data["value_fish_ht"] = fish_purchases.aggregate(total=Sum("price_ht"))["total"] or 0
 
-    fish_egalim = fish_purchases.filter(characteristics__overlap=EGALIM_LABELS)
+    fish_egalim = fish_purchases.filter(characteristics__overlap=PURCHASE_EGALIM_LABELS)
     data["value_fish_egalim_ht"] = fish_egalim.aggregate(total=Sum("price_ht"))["total"] or 0
